refactor: centralize Decky-internal global reaches in one module
